# Conflicts:
#	app/src/main/AndroidManifest.xml
#	local.properties

+++ app/src/main/java/ca/georgebrown/comp3074/project/Event/Event.java
@@ -26,7 +26,7 @@ public class Event implements Serializable {
         return Event_Id;
     }
 
-    public void setEvent_Id(int event_Id) {
+    public void setEvent_Id(long event_Id) {
         Event_Id = event_Id;
     }
